scrapping backend/requirements.txt
selenium
chromedriver-py
beautifulsoup4
requests
readability-lxml
pandas